src/main/java/com/product/sell/sell/service/impl/ProductCategoryServiceImpl.java
package com.product.sell.sell.service.impl;

import com.product.sell.sell.dao.ProductCategory;
import com.product.sell.sell.repositories.ProductCategoryRepository;
import com.product.sell.sell.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryRepository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryIn(List<Integer> categories) {
        return productCategoryRepository.findByCategoryTypeIn(categories);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
